+++ autumn/projects/sm_sir/bangladesh/bangladesh/project.py
@@ -1,9 +1,7 @@
 import json
 from datetime import date
-from typing import Union
-import numpy as np
-import pandas as pd
 
+from autumn.tools.utils.utils import wrap_function_for_series
 from autumn.models.sm_sir.parameters import BASE_DATE
 from autumn.tools.project import Project, ParameterSet, load_timeseries, build_rel_path
 from autumn.tools.calibration import Calibration
@@ -33,49 +31,7 @@ late_deaths = ts_set["infection_deaths"].loc[notifications_trunc_point:]
 
 
 def get_diff(series):
-    return series.diff().rolling(window=7).mean()
-
-
-def wrap_function_for_series(
-        process_to_apply: callable
-) -> callable:
-    """
-    Apply a function to a pandas series that can cope with the series coming in either directly as pandas or as a numpy
-    array.
-
-    Args:
-        process_to_apply: A function that can be applied to a pandas series
-
-    Returns:
-        The processed series
-
-    """
-
-    def apply_function_to_series(
-            input_series: Union[pd.Series, np.ndarray]
-    ) -> Union[pd.Series, np.ndarray]:
-        """
-        The function that can be applied directly to the series data.
-
-        Args:
-            input_series: Pandas or numpy array, either the input timeseries or the equivalent model derived value
-
-        Returns:
-            Function that can be applied to either pandas or numpy
-
-        """
-
-        # Find the input format
-        is_numpy = isinstance(input_series, np.ndarray)
-        working_series = pd.Series(input_series) if is_numpy else input_series
-
-        # The actual manipulation to the series
-        working_series = process_to_apply(working_series)
-
-        # Return in the appropriate format
-        return working_series.to_numpy() if is_numpy else working_series
-
-    return apply_function_to_series
+    return series.diff().rolling(window=50).mean()
 
 
 processed_output = "notif_change"
